Einige Warnungen wegen 64bit-site_t bearbeitet
git-svn-id: http://subversion.sos/svn/sos/scheduler/engine@17138 c11b8382-0eeb-0310-9acf-d3b3cacd313b

+++ opensource/kernel-cpp/kram/log.cxx
@@ -557,7 +557,7 @@ void Mswin_debug_streambuf::get_app_name()
                 while( p > path  &&  p[-1] != '/'  &&  p[-1] != '\\'  &&  p[-1] != ':' )  p--;
                 while( q > p  &&  q[-1] != '.' )  q--;
                 if( q > p  &&  q[-1] == '.' ) q--;
-                int l = min( sizeof _app_name - 1, uint( q - p ) );
+                int l = min( sizeof _app_name - 1, size_t( q - p ) );
                 memcpy( _app_name, p, l );
                 _app_name[ l ] = '\0';
                 //sprintf( _app_name, "%04X ", (uint)htask );
